+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 69.0.0
 
 * BREAKING: remove redundant test helpers for finding subscriber lists
 * BREAKING: Switch "find_or_create_subscriber_list" to only call "create" endpoint